+++ plugins/security/api_roles.go
@@ -69,18 +69,3 @@ func (c rolesClient) Patch(ctx context.Context, req RolesPatchReq) (RolesPatchRe
 
 	return data, nil
 }
-
-// RolesIndexPermission contains index permissions and is used for Get and Put requests
-type RolesIndexPermission struct {
-	IndexPatterns  []string `json:"index_patterns,omitempty"`
-	DLS            string   `json:"dls,omitempty"`
-	FLS            []string `json:"fls,omitempty"`
-	MaskedFields   []string `json:"masked_fields,omitempty"`
-	AllowedActions []string `json:"allowed_actions,omitempty"`
-}
-
-// RolesTenantPermission contains tenant permissions and is used for Get and Put requests
-type RolesTenantPermission struct {
-	TenantPatterns []string `json:"tenant_patterns,omitempty"`
-	AllowedActions []string `json:"allowed_actions,omitempty"`
-}